Added files via upload
not complete!! having issues with getting keys to turn color. right now, i have a key turning red when pressed. i put in a delay of half a second, so ideally the key would light up and then turn back to white or black. instead, it just stays red unless the mouse moves over the window (even if nothing is pressed). this is weird and funky and i am confused.